+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog is loose. Versions are not semantic, they are incremental. Splint
 
 ## Unreleased
 
+### New rules
+
+- `lint/redundant-call-str`: Don't call `str` on input that's guaranteed to be a string: Prefer `"foo"` to `(str "foo")`, `(str "foo" bar)` to `(str (str "foo" bar))`, and `(format "foo%s" bar)` to `(str (format "foo%s" bar))`. (See [clj-kondo#2323](https://github.com/clj-kondo/clj-kondo/issues/2323) for inspiration.)
+
 ### Changed
 
 - Switched from `clojure.pprint` to [fipp](https://github.com/brandonbloom/fipp) for pretty-printing code. Fast and easy to extend.